move runMigration to correct place
